% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{as_argument}
\alias{as_argument}
\title{Coercion to an argument type}
\usage{
as_argument(x)
}
\arguments{
\item{x}{value to be coerced}
}
\value{
An argument list object
}
\description{
Coercion to an argument type
}
\examples{

a <- as_argument("numeric")
b <- as_argument(~ is.logical(.) && .)
c <- as_argument(list(1, 2, 3, TRUE, FALSE))

}
