% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicate.R
\name{as_predicate}
\alias{as_predicate}
\title{Convert an argument check onto a predicate}
\usage{
as_predicate(x)
}
\arguments{
\item{x}{Generally some argument type}
}
\value{
A predicate (unary function which returns a logical value)
}
\description{
Convert an argument check onto a predicate
}
\examples{

a <- as_argument("numeric")[[1]]
as_predicate(a)(10) # => TRUE
as_predicate(a)('10') # => FALSE

}
