% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{def}
\alias{def}
\title{Define a function with argument checking}
\usage{
def(.f, ...)
}
\arguments{
\item{.f}{Function}

\item{...}{Named arguments for the function}
}
\value{
A function that will check its arguments when called
}
\description{
`def` is the main utility of this package. It allows users to define
functions that perform argument checking when called.
}
\examples{

# Check argument type
succ <- def(x = "numeric", {
  x + 1
})

succ(1) # => 2
\donttest{
succ('a') # => error
}


# Check if argument is in a set of possible values
f1 <- def(x = list(TRUE, FALSE, NA, NULL), {
  x && !x
})

# Check any condition on an argument
f2 <- def(x = (~ . > 0 && . < 10), {
  x + 1
})

# Chain multiple checks
f3 <- def(x = "character" \%&\% (~ length(grep("hi", .)) > 0), {
  paste(x, "world!")
})

# Set default values
f4 <- def(x = "numeric" \%=\% 0, {
  rep(TRUE, x)
})

}
