% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{defaults_to}
\alias{defaults_to}
\alias{\%=\%}
\title{Set a default value for an argument}
\usage{
defaults_to(value)

alist \%=\% value
}
\arguments{
\item{value}{Default value}

\item{alist}{Left-hand side, argument definitions such as classes or conditions}
}
\description{
Set a default value for an argument
}
\examples{

f1 <- def(x = defaults_to(0), { x + 1 })
f2 <- def(x = "numeric" \%=\% 0, { x + 1 })

f1() # => 1
f2() # => 1

}
