% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argcheck.R
\name{example_argcheck}
\alias{example_argcheck}
\title{Example function}
\usage{
example_argcheck(number = NULL, string = NULL)
}
\arguments{
\item{number}{A number}

\item{string}{A string}
}
\value{
dunno
}
\description{
Example function
}
