% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{\%&\%}
\alias{\%&\%}
\title{Chain multiple argument conditions}
\usage{
a1 \%&\% a2
}
\arguments{
\item{a1}{First condition}

\item{a2}{Second condition}
}
\value{
The concatenation of both conditions
}
\description{
Chain multiple argument conditions
}
\examples{

checklist <-
  "numeric" \%&\%
  (~ . != 0) \%&\%
  (~ abs(.) < 10)

f <- def(a = "numeric", b = checklist, { a / b })

}
